:recycle:Refactor: split Header component 헤더, 알림, 알림 모달 컴포넌트 분리
