+++ tradin/src/components/commons/header/main/Header.types.ts
@@ -1,0 +1,7 @@
+export interface HeaderUIProps {
+    handleTitleClick: () => void;
+    currentAnnouncement: number;
+    onClickSignOut: () => void;
+    onClickMoveToProfile: () => void;
+    onClickMoveToLogin: () => void;
+}
