+++ tradin/src/components/commons/header/main/MockAnnouncements.tsx
@@ -1,0 +1,7 @@
+export const announcements = [
+    { title: "공시 1", content: "공시 1에 대한 내용" },
+    { title: "공시 2", content: "공시 2에 대한 내용" },
+    { title: "공시 3", content: "공시 3에 대한 내용" },
+    { title: "공시 4", content: "공시 4에 대한 내용" },
+    { title: "공시 5", content: "공시 5에 대한 내용" },
+];
